# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/scm/purchase-orders": {
      "post": {
        "operationId": "PurchaseOrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PO created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new Purchase Order",
        "tags": [
          "scm"
        ]
      },
      "get": {
        "operationId": "PurchaseOrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all Purchase Orders",
        "tags": [
          "scm"
        ]
      }
    },
    "/scm/purchase-orders/{id}": {
      "get": {
        "operationId": "PurchaseOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single Purchase Order by ID",
        "tags": [
          "scm"
        ]
      }
    },
    "/scm/purchase-orders/{id}/down-payment": {
      "post": {
        "operationId": "PurchaseOrdersController_createDP",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Down Payment for a PO",
        "tags": [
          "scm"
        ]
      }
    },
...
    "/scm/purchase-returns": {
      "post": {
        "operationId": "PurchaseReturnsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseReturns"
        ]
      },
      "get": {
        "operationId": "PurchaseReturnsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
    "/scm/purchase-returns/{id}": {
      "get": {
        "operationId": "PurchaseReturnsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
    "/scm/purchase-returns/{id}/status": {
      "patch": {
        "operationId": "PurchaseReturnsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseReturnStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PurchaseReturns"
        ]
      }
    },
...
    "/rnd/samples": {
      "post": {
        "operationId": "RndController_createSample",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSampleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new sample request",
        "tags": [
          "rnd"
        ]
      },
      "get": {
        "operationId": "RndController_getSamples",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/sample/{id}/advance": {
      "patch": {
        "operationId": "RndController_advanceSample",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceSampleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/sample/{id}/accept": {
      "post": {
        "operationId": "RndController_acceptSample",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/dashboard": {
      "get": {
        "operationId": "RndController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get R&D dashboard metrics",
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/samples/{id}": {
      "get": {
        "operationId": "RndController_getSample",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/inbox": {
      "get": {
        "operationId": "RndController_getInbox",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/staffs": {
      "get": {
        "operationId": "RndController_getStaffs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/sample/{id}/assign": {
      "patch": {
        "operationId": "RndController_assignPIC",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/samples/{id}/versions": {
      "get": {
        "operationId": "RndController_getVersions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/samples/{id}/feedback": {
      "get": {
        "operationId": "RndController_getFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/lab-test-results": {
      "get": {
        "operationId": "RndController_getAllLabTestResults",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all lab test results (supports ?type=stability)",
        "tags": [
          "rnd"
        ]
      },
      "post": {
        "operationId": "RndController_createLabTestResult",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
    "/rnd/lab-test-results/{formulaId}": {
      "get": {
        "operationId": "RndController_getLabTestResults",
        "parameters": [
          {
            "name": "formulaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rnd"
        ]
      }
    },
...
  },
  "info": {
    "title": "NexERP API",
    "description": "The ultimate manufacturing ERP system API documentation",
    "version": "4.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PO-2024-001",
            "description": "Human-readable PO ID"
          },
          "supplierId": {
            "type": "string",
            "format": "uuid",
            "example": "uuid-supplier-123"
          },
          "leadId": {
            "type": "string",
            "format": "uuid",
            "example": "uuid-lead-456"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "totalAmount": {
            "type": "number",
            "example": 1500000
          },
          "estArrival": {
            "type": "string",
            "example": "2024-12-31"
          },
          "escalationPin": {
            "type": "string"
          },
          "escalationReason": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "supplierId"
        ]
      },
...
      "CreatePurchaseReturnItemDto": {
        "type": "object",
        "properties": {
          "materialId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          }
        },
        "required": [
          "materialId",
          "quantity",
          "unitPrice"
        ]
      },
      "CreatePurchaseReturnDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "inboundId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseReturnItemDto"
            }
          }
        },
        "required": [
          "supplierId",
          "warehouseId",
          "items"
        ]
      },
      "UpdatePurchaseReturnStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "object"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateSampleDto": {
        "type": "object",
        "properties": {
          "leadId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string",
            "example": "Luminous Serum"
          },
          "targetFunction": {
            "type": "string",
            "example": "Brightening"
          },
          "textureReq": {
            "type": "string",
            "example": "Liquid Gel"
          },
          "colorReq": {
            "type": "string",
            "example": "Clear"
          },
          "aromaReq": {
            "type": "string",
            "example": "Floral"
          },
          "targetHpp": {
            "type": "number",
            "example": 15000
          },
          "targetDeadline": {
            "type": "string",
            "example": "2026-12-31T23:59:59Z"
          },
          "difficultyLevel": {
            "type": "number",
            "example": 1
          },
          "picId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "leadId",
          "productName",
          "targetFunction",
          "textureReq",
          "colorReq",
          "aromaReq"
        ]
      },
...
      "AdvanceSampleDto": {
        "type": "object",
        "properties": {
          "newStage": {
            "type": "object"
          },
          "feedback": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          }
        },
        "required": [
          "newStage"
        ]
      },
...
    }
  }
}